django_project/requirements.txt
certifi==2018.10.15
chardet==3.0.4
Django==2.1.3
django-crispy-forms==1.7.2
idna==2.7
Pillow==5.4.0
pygame==1.9.3
python-dateutil==2.7.5
pytz==2018.7
requests==2.20.1
six==1.11.0
urllib3==1.24.1
watson-developer-cloud==2.4.1
websocket-client==0.47.0
